requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.22
bleach==1.4.2
celery==3.1.20
conceptpower-api==1.1
diff-match-patch==20121119
dj-database-url==0.4.0
Django==1.9.2
django-crispy-forms==1.6.0
django-filter==0.13.0
django-grappelli==2.7.3
django-gunicorn==0.1.1
django-haystack==2.5.dev0
django-markupfield==1.4.0
django-reversion==1.10.1
djangorestframework==3.3.3
gunicorn==19.4.5
html5lib==0.9999999
isodate==0.5.4
keepalive==0.5
kombu==3.0.33
Markdown==2.6.5
psycopg2==2.6.1
pyparsing==2.1.0
pytz==2015.7
rdflib==4.2.1
redis==2.10.5
requests==2.9.1
six==1.10.0
SPARQLWrapper==1.7.6
wheel==0.24.0
whitenoise==2.0.6
Whoosh==2.7.0
